pyproject.toml
[project]
name = "clauded"
version = "0.1.0"
description = "Isolated, per-project Lima VMs with automatic environment provisioning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "questionary>=2.0",
    "pyyaml>=6.0",
    "ansible>=13.2",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
clauded = "clauded.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-pyyaml>=6.0",
    "pre-commit>=4.0",
    "hypothesis>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clauded"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
    "src/clauded/**/*.py",
    "src/clauded/roles/**/*.yml",
    "src/clauded/linguist/**/*.yml",
    "src/clauded/linguist/README.md",
]

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

# Mypy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["questionary.*", "ansible.*"]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
